Move leadershipBoard layout to views directory
