+++ src/models/flappyBirdGame.js
@@ -10,7 +10,7 @@ const flappyBirdGame = () => {
     dom: {
       createContainer: true,
     },
-    scene: [Playable, GameOver],
+    scene: [GameOver],
     scale: {
       mode: Phaser.Scale.FIT,
       autoCenter: Phaser.Scale.CENTER_BOTH,
